+++ src/main/java/com/foodielog/server/application/feed/entity/FeedLike.java
@@ -1,11 +1,8 @@
-package com.foodielog.server.application.notification;
+package com.foodielog.server.application.feed.entity;
 
 import java.sql.Timestamp;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -13,41 +10,36 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.CreationTimestamp;
 
-import com.foodielog.server.application.notification.type.NotificationType;
+import com.foodielog.server.application.notification.entity.Notification;
 import com.foodielog.server.application.user.entity.User;
-import com.foodielog.server.application.user.type.Flag;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "notification_tb")
 @Getter
-public class Notification {
+@Table(name = "feed_like_tb")
+@Entity
+public class FeedLike {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne
+	@JoinColumn(name = "feed_id")
+	private Feed feed;
+
+	@ManyToOne
 	@JoinColumn(name = "user_id")
 	private User user;
 
-	@Column(nullable = false)
-	private NotificationType type;
-
-	@Column(nullable = false)
-	private Long contentId;
-
-	@Column(nullable = false, length = 1)
-	@Enumerated(EnumType.STRING)
-	@ColumnDefault("'N'")
-	private Flag checkFlag;
+	@ManyToOne
+	@JoinColumn(name = "notification")
+	private Notification notification;
 
 	@CreationTimestamp
-	private Timestamp createdAt;
+	private Timestamp createAt;
 }
