+++ src/main/java/com/foodielog/server/application/notification/entity/Notification.java
@@ -1,8 +1,11 @@
-package com.foodielog.server.application.feed;
+package com.foodielog.server.application.notification.entity;
 
 import java.sql.Timestamp;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -10,36 +13,41 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
+import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.CreationTimestamp;
 
-import com.foodielog.server.application.notification.Notification;
+import com.foodielog.server.application.notification.type.NotificationType;
 import com.foodielog.server.application.user.entity.User;
+import com.foodielog.server.application.user.type.Flag;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "feed_like_tb")
 @Entity
-public class FeedLike {
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "notification_tb")
+@Getter
+public class Notification {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne
-	@JoinColumn(name = "feed_id")
-	private Feed feed;
-
-	@ManyToOne
 	@JoinColumn(name = "user_id")
 	private User user;
 
-	@ManyToOne
-	@JoinColumn(name = "notification")
-	private Notification notification;
+	@Column(nullable = false)
+	private NotificationType type;
+
+	@Column(nullable = false)
+	private Long contentId;
+
+	@Column(nullable = false, length = 1)
+	@Enumerated(EnumType.STRING)
+	@ColumnDefault("'N'")
+	private Flag checkFlag;
 
 	@CreationTimestamp
-	private Timestamp createAt;
+	private Timestamp createdAt;
 }
