+++ src/main/java/com/foodielog/server/application/restaurant/entity/Restaurant.java
@@ -1,10 +1,15 @@
-package com.foodielog.server.application.restaurant;
+package com.foodielog.server.application.restaurant.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
